Move States into a new states.pyx
