+++ dwave/optimization/model.pyi
@@ -112,33 +112,6 @@ class Model:
     def unlock(self): ...
 
 
-class States:
-    def __init__(self, model: Model): ...
-    def __len__(self) -> int: ...
-    def clear(self): ...
-
-    def from_file(
-        self,
-        file: typing.Union[typing.BinaryIO, collections.abc.ByteString, str],
-        *,
-        replace: bool = True,
-        check_header: bool = True,
-        ) -> Model: ...
-
-    def from_future(self, future: object, result_hook: collections.abc.Callable): ...
-    def initialize(self): ...
-
-    def into_file(
-        self,
-        file: typing.Union[typing.BinaryIO, collections.abc.ByteString, str],
-        ): ...
-
-    def resize(self, n: int): ...
-    def resolve(self): ...
-    def size(self) -> int: ...
-    def to_file(self) -> typing.BinaryIO: ...
-
-
 class Symbol:
     def __init__(self, *args, **kwargs) -> typing.NoReturn: ...
     def equals(self, other: Symbol) -> bool: ...
